Change input name for the Slack sign-up
The input tag in the Slack sign-up form had the wrong name. The name was "name" rather than "email" and the Express route is looking for the name "email". As a result, the if statement in the route is failing because req.body.email doesn't exist. The HTML is providing req.body.name. Rather than change the Express route I just renamed the name of the input tag. I changed it to "email" using the Chrome Developer Tool (too lazy to do a full clone using git) and everything worked.